separe devtools interfaces from main file
